requirements.txt
torch>=1.3
opencv-python==4.5.5.64
opencv-contrib-python
decord >= 0.4.1
Pillow==8.2.0
mmcv==1.3.6
pandas==1.1.4
scipy==1.6.3
einops==0.4.1
matplotlib==3.4.2
numpy==1.21.1